Add - is file? in monitoring system

+++ main.py
@@ -227,28 +227,28 @@ class MyFileHandler(FileSystemEventHandler):
         
       print('file: ' + file)   
       print('path: ' + path)
-      
-      if folder.hashdecision == 'ON':
-        print("convert " + file + " to md5")
-        file = renameToMd5(path + '/', file)
-        time.sleep(1)
-        print("New name is:" + file)
-      else:
-        print("Hash decision off")
-      print("\nLet's check if the " + file + " already exists in the log\n")
-      if checkLog(folder.path, folder.folder_name, folder.log_name, file) == False:
-        print("It's a new file!")
-        if sendToTelegram(folder.path, folder.folder_name, file, folder.chat_id, tokenBot) == False:
-          print("File not send")
+      if os.path.isfile(path + '/' + file) == True:
+        if folder.hashdecision == 'ON':
+          print("convert " + file + " to md5")
+          file = renameToMd5(path + '/', file)
+          time.sleep(1)
+          print("New name is:" + file)
         else:
-          print("We don't have the file in the log")
-          writeInLog(folder.path, folder.log_name, file)
-          os.remove(folder.path + folder.folder_name + '/' + file)
-          time.sleep(3)
-
-
+          print("Hash decision off")
+        print("\nLet's check if the " + file + " already exists in the log\n")
+        if checkLog(folder.path, folder.folder_name, folder.log_name, file) == False:
+          print("It's a new file!")
+          if sendToTelegram(folder.path, folder.folder_name, file, folder.chat_id, tokenBot) == False:
+            print("File not send")
+          else:
+            print("We don't have the file in the log")
+            writeInLog(folder.path, folder.log_name, file)
+            os.remove(folder.path + folder.folder_name + '/' + file)
+            time.sleep(3)
+        else:
+          print("Not send.. file already exist in log")
       else:
-        print("Not send.. file already exist in log")
+        print("Event isn't a file...")
